requirements.txt
discord.py>=2.3
PlexAPI>=4.15
aiosqlite>=0.19
python-dotenv>=1.0
